organize files in config directory
